Refactor API to use `int` instead of `uint` for prefix lengths
By convention of Go, lengths and indexes are always `int`.

+++ internal/utils/utils.go
@@ -10,11 +10,11 @@ import "github.com/nextmn/rfc9433/encoding/errors"
 // ipv6: Address to extract bits from
 // startBit: offset in bits
 // length: length of result in Bytes
-func FromIPv6(ipv6 [16]byte, startBit uint, length uint) ([]byte, error) {
-	if uint(len(ipv6)) < length {
+func FromIPv6(ipv6 [16]byte, startBit int, length int) ([]byte, error) {
+	if len(ipv6) < length {
 		return nil, errors.ErrTooShortToParse
 	}
-	if startBit+uint(length*8) > 8*uint(len(ipv6)) {
+	if startBit+(length*8) > 8*len(ipv6) {
 		return nil, errors.ErrOutOfRange
 	}
 	startByte := startBit / 8
@@ -38,15 +38,16 @@ func FromIPv6(ipv6 [16]byte, startBit uint, length uint) ([]byte, error) {
 
 // usage conditions :
 // 1. slice must be large enough
-// 2. every bit after endBit should be zero (no reset is performed in the function)
-func AppendToSlice(slice []byte, endBit uint, appendThis []byte) error {
+// 2. endBit must be positive
+// 3. every bit after endBit should be zero (no reset is performed in the function)
+func AppendToSlice(slice []byte, endBit int, appendThis []byte) error {
 	endByte := endBit / 8
 	offset := endBit % 8
 	isOffset := 0
 	if offset > 0 {
 		isOffset = 1
 	}
-	if isOffset+int(endByte)+len(appendThis) > len(slice) {
+	if isOffset+endByte+len(appendThis) > len(slice) {
 		return errors.ErrTooShortToMarshal
 	}
 	if offset == 0 {
@@ -56,11 +57,11 @@ func AppendToSlice(slice []byte, endBit uint, appendThis []byte) error {
 	}
 	//  add right part of bytes
 	for i, b := range appendThis {
-		slice[int(endByte)+i] |= b >> offset
+		slice[endByte+i] |= b >> offset
 	}
 	// add left part of bytes
 	for i, b := range appendThis {
-		slice[int(endByte)+isOffset+i] |= b << (8 - offset)
+		slice[endByte+isOffset+i] |= b << (8 - offset)
 	}
 	return nil
 }
